+++ VSIX/HeleprsCs.Visual/Controls/Separator.xaml.cs
