+++ VSIX/HeleprsCs.Visual/Helpers/BaseUserControl.cs
