+++ VSIX/HeleprsCs.Visual/Helpers/BindingProxy.cs
