+++ VSIX/HeleprsCs.Visual/Helpers/FocusWatcher.cs
