+++ VSIX/HeleprsCs.Visual/Helpers/Graphics.cs
