+++ VSIX/HeleprsCs.Visual/Helpers/MenuItem.cs
