+++ VSIX/HeleprsCs.Visual/Properties/InvalidatableProperty.cs
