+++ VSIX/HeleprsCs.Visual/Properties/VisibilityProperty.cs
