+++ VSIX/HeleprsCs.Visual/Time/TimerManager.cs
