+++ VSIX/HelpersCs.Attributes/Attributes/FieldAttributes.cs
@@ -15,6 +15,8 @@
 
     [System.AttributeUsage(System.AttributeTargets.Field)]
     public sealed class ObservablePropertyAttribute : System.Attribute {
+        public string? NotifyMethod { get; set; }
+
         public ObservablePropertyAttribute(
             Markers.Access.Get p0 = default,
             Markers.Access.Set p1 = default
@@ -23,6 +25,18 @@
         public ObservablePropertyAttribute(
             Markers.Access.Get p0,
             Markers.Access.PrivateSet p1
+            ) {
+        }
+    }
+
+
+    [System.AttributeUsage(System.AttributeTargets.Field)]
+    public sealed class ObservableMultiStatePropertyAttribute : System.Attribute {
+        public string? NotifyMethod { get; set; }
+
+        public ObservableMultiStatePropertyAttribute(
+            Markers.Access.Get p0 = default,
+            Markers.Access.PrivateSet p1 = default
             ) {
         }
     }
@@ -44,10 +58,6 @@
             Markers.Access.Get p0
             ) {
         }
-        //public InvalidatablePropertyAttribute(
-        //    Markers.Access.PrivateSet p0
-        //    ) {
-        //}
     }
