+++ VSIX/HelpersCs/Helpers/Events/TriggeredAction.cs
@@ -1,16 +1,4 @@
-﻿using System;
-using System.Linq;
-using System.Text;
-using System.Windows.Threading;
-using System.Runtime.CompilerServices;
-using System.Threading;
-using System.Threading.Tasks;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Collections.Specialized;
-using System.ComponentModel;
-
-namespace Helpers {
+﻿namespace Helpers {
     namespace Events {
         public class TriggeredAction {
             private bool _isTriggered = false;
