+++ VSIX/HelpersCs/Helpers/RepresentationsTableBase.cs
@@ -3,7 +3,6 @@ using System.Linq;
 using System.Text;
 using System.Windows;
 using System.Windows.Input;
-using System.Windows.Threading;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
@@ -11,6 +10,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Collections.Specialized;
 using System.ComponentModel;
+
 
 namespace Helpers {
     public abstract class RepresentationsTableBase<TRecord> {
